csd: add a frame window
To improve the compatibility of client-side decorations with
traditional decorations, we introduce a frame window, which is
the actual toplevel. widget->window for the GtkWindow is a child
of this window, and just as with traditional window decorations,
corresponds to the window content.

All sizes (default size, size allocation, parameters given
to gtk_window_resize, etc) refer to the size of the content.

The resize handle windows and the titlebar widgets are children
of the frame window.